# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "chat-API",
      "description": "A little whatsapp api clone",
      "version": "1.0"
    },
    "paths": {
      "/api/v1/users": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "register",
          "operationId": "register",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/registerrequest"
                    },
                    {
                      "example": {
                        "firstname": "Cam",
                        "lastname": "Nort",
                        "email": "Canort@gmail.com",
                        "password": "Ganar01234@."
                      }
                    }
                  ]
                },
                "example": {
                  "firstname": "Cam",
                  "lastname": "Nort",
                  "email": "Canort@gmail.com",
                  "password": "Ganar01234@."
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false
        },
        "get": {
          "tags": [
            "users"
          ],
          "summary": "get all users",
          "operationId": "getallusers",
          "parameters": [],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
      "/api/v1/messages/conversation/{id}": {
        "post": {
          "tags": [
            "messages"
          ],
          "summary": "create new message",
          "operationId": "createnewmessage",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 5
              }
            }
          ],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/createnewmessagerequest"
                    },
                    {
                      "example": {
                        "content": "hola numero uno con autenticacion",
                        "senderId": 6
                      }
                    }
                  ]
                },
                "example": {
                  "content": "hola numero uno con autenticacion",
                  "senderId": 6
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "get": {
          "tags": [
            "messages"
          ],
          "summary": "get conversation messages",
          "operationId": "getconversationmessages",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32",
                "example": 5
              }
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": [
            {
              "bearer": []
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "registerrequest": {
          "title": "registerrequest",
          "required": [
            "firstname",
            "lastname",
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "firstname": {
              "type": "string"
            },
            "lastname": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "firstname": "Cam",
            "lastname": "Nort",
            "email": "Canort@gmail.com",
            "password": "Ganar01234@."
          }
        },
...
        "createnewmessagerequest": {
          "title": "createnewmessagerequest",
          "required": [
            "content",
            "senderId"
          ],
          "type": "object",
          "properties": {
            "content": {
              "type": "string"
            },
            "senderId": {
              "type": "integer",
              "format": "int32"
            }
          },
          "example": {
            "content": "hola numero uno con autenticacion",
            "senderId": 6
          }
        }
      },
      "securitySchemes": {
        "bearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
    "security": [],
...
  }